+++ src/adapters/moonpay/index.js
@@ -1,1 +1,0 @@
-export { moonpayAdapter } from './client.js';